feat(view): 興味を尋ねるページをtamaguiに対応
