+++ src/constant/meta.tsx
@@ -1,23 +1,27 @@
 import { TFunction } from "i18next";
+import { useAppTranslation } from "src/hooks/useAppTranslation";
 
-export const PageMetaData = (t: TFunction) => ({
-    top: {
-        title: t("ogp:topPageTitle"),
-        description: t("ogp:topPageDescription"),
-    },
-    place: {
-        search: {
-            title: t("ogp:placeSearchPageTitle"),
-            description: t("ogp:placeSearchPageDescription"),
+export const PageMetaData = (translation?: TFunction) => {
+    const t = translation || useAppTranslation().t;
+    return {
+        top: {
+            title: t("ogp:topPageTitle"),
+            description: t("ogp:topPageDescription"),
         },
-    },
-    plans: {
-        interest: {
-            title: (fromCurrentLocation: boolean) =>
-                fromCurrentLocation
-                    ? t("ogp:planInterestPageFromCurrentLocationTitle")
-                    : t("ogp:planInterestPageFromSelectedPlaceTitle"),
-            description: t("ogp:planInterestPageDescription"),
+        place: {
+            search: {
+                title: t("ogp:placeSearchPageTitle"),
+                description: t("ogp:placeSearchPageDescription"),
+            },
         },
-    },
-});
+        plans: {
+            interest: {
+                title: (fromCurrentLocation: boolean) =>
+                    fromCurrentLocation
+                        ? t("ogp:planInterestPageFromCurrentLocationTitle")
+                        : t("ogp:planInterestPageFromSelectedPlaceTitle"),
+                description: t("ogp:planInterestPageDescription"),
+            },
+        },
+    };
+};
