+++ src/view/plan/MatchInterestPageTemplate.tsx
@@ -1,6 +1,7 @@
-import { VStack } from "@chakra-ui/react";
 import { ReactNode } from "react";
-import { MessageCard } from "src/view/plan/MessageCard";
+import { Padding } from "src/constant/padding";
+import { Size } from "src/constant/size";
+import { Text, YStack } from "tamagui";
 
 type Props = {
     message: string;
@@ -14,20 +15,40 @@ export function MatchInterestPageTemplate({
     children,
 }: Props) {
     return (
-        <VStack h="100%" w="100%" spacing={0}>
+        <YStack h="100%" w="100%" alignItems="center" gap={0}>
             {navBar}
-            <VStack
+            <YStack
                 flex={1}
                 h="100%"
                 w="100%"
-                maxWidth="990px"
-                px="16px"
-                pt="8px"
-                pb={{ base: "64px", sm: "32px" }}
+                maxWidth={Size.mainContentWidth}
+                px={Padding.p16}
+                pt={Padding.p8}
+                gap={Padding.p8}
+                $sm={{
+                    pb: Padding.p32,
+                }}
+                $gtSm={{
+                    pb: Padding.p64,
+                }}
             >
-                <MessageCard message={message} />
-                {children}
-            </VStack>
-        </VStack>
+                <YStack
+                    alignItems="center"
+                    justifyContent="center"
+                    width="100%"
+                    px={Padding.p16}
+                    py={Padding.p32}
+                    borderWidth={1}
+                    borderRadius={5}
+                    borderColor="rgba(0, 0, 0, 0.1)"
+                    backgroundColor="white"
+                >
+                    <Text userSelect="none">{message}</Text>
+                </YStack>
+                <YStack w="100%" flex={1}>
+                    {children}
+                </YStack>
+            </YStack>
+        </YStack>
     );
 }
